==> scripts/1_getStudy1dataQualtrics.Rmd <==
---
title: "read Study 1 qualtrics data"
author: "Jen Richmond"
date: "12/03/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# load packages + set API

The qualtRics package reads data straight from qualtrics into R via the API and saves the dataset as a .csv in the data folder. 

https://cran.r-project.org/web/packages/qualtRics/vignettes/qualtRics.html


```{r message=FALSE, warning=FALSE}
library(qualtRics)
library(tidyverse)
library(janitor)
library(here)

qualtrics_api_credentials(api_key =     "Fo3JT3EY4XvmrZUh0Y2ELmQUig66DM9I9PNROSbz",
                          base_url = "unsw.syd1.qualtrics.com",
                          install = TRUE, 
                          overwrite = TRUE)


# restart R or set enviornment 

readRenviron("~/.Renviron")
```


JennyC surveyID = SV_d4PE06XiLBkNmse

```{r}
questions <- survey_questions(surveyID = "SV_d4PE06XiLBkNmse")
```



```{r}
data1 <- fetch_survey(surveyID = "SV_d4PE06XiLBkNmse", 
                         force_request = TRUE, 
                         verbose = TRUE) %>%
  clean_names() 

currentDate <- Sys.Date()

csvFileName <- paste(currentDate,"_Study1data.csv",sep="")

data1 %>% write_csv(here("data", file=csvFileName))




```


Pulling out identifiable details for people who want to hear about results 
```{r}

consent_results <- data %>%
  select(name, address, email_address, consent) 

consent_results %>%
  write_csv(here("data", "identifiable_details", "people_who_want_results.csv" ))

```

# PULL JUST A SECTION TO PLAY WITH 

```{r}
datatest20 <- sample_n(data, 20)

datatest20 %>%
  write_csv(here("data", "N20jennytest.csv"))

```
