==> scripts/2_cleanStudy1_Qualtricsdata.Rmd <==
---
title: "cleanQualtricsdata"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r message=FALSE, warning=FALSE}
library(tidyverse)
library(here)
library(janitor)
```

```{r message=FALSE, warning=FALSE}
datatest <-  read_csv(here("data", "2021-07-13_Study1data.csv"))
```

# select just variables of interest

Make a dataframe with just useful variables

Selecting just participant details, ratings of perception, attitude, intention, manipulation checks and attention variables. Lets keep attention and other variables together for now.

```{r}

data_select <- datatest %>%
  select(end_date, progress, response_id, group, gender, age_in_years, faculty_of_study, research_experience, perception_1, starts_with("attitude"), starts_with("intention"), manipulation_check_gf_1, manipulation_check_lf_1, starts_with("attention_attempt"),  starts_with("score"))

```

# filter out participants who didn't complete the whole survey

Looks like there are some value of < 100 in the progress column. Filter out those that didn't complete the survey. 

Goes from 175 to 161 participants = 14 participants didn't finish

```{r}

progress_filtered <- data_select %>%
  filter(progress == "100")

progress_filtered %>% 
  count()

```

# create new exclude column for attention scores 

I think it is easier to do this without making the data long...

use mutate and case_when to create a new column that puts "include" or "exclude" in the column, depending on performance on score_attempt 1, 2 and 3. If they get >-5 on 1, 2, or, 3 then 'include', if they get <=4 on attempt 3 then 'exclude'

http://jenrichmond.rbind.io/post/mutate-and-if-else-to-create-new-variables/ 

```{r}

exclude_column <- progress_filtered %>%
  mutate(atten_exclusion = case_when(score_attempt_1 >= 5 ~ 'Include',
                               score_attempt_2 >= 5 ~ 'Include',
                               score_attempt_3 >= 5 ~ 'Include',
                               TRUE ~ 'Exclude'))

```

# filter out those that get exclude in above column 

Filter out those excluded for attention fail, then drop atten/score columns from the dataframe using select(). 

Goes from 161 to 145 participants = 16 participants fail atten checks

```{r}

attention_filtered <- exclude_column %>%
  filter(atten_exclusion == "Include")

important_data <- attention_filtered %>%
  select(end_date, progress, response_id, group, gender, age_in_years, faculty_of_study, research_experience, perception_1, starts_with("attitude"), starts_with("intention"), manipulation_check_gf_1, manipulation_check_lf_1, atten_exclusion)

```

#  exclude participants who didn't answer enough attitude/intention qs 

from prereg: If a participant does not complete the single perceived severity question, at least 3 of 5 attitude questions, or at least 2 of 4 implementation intention questions, the participant’s response will be excluded from the analysis for that particular dependent variable. 

Need to count how many NAs each participant has across perception, attitude and intention qs and exclude if more than 0 (perception) or 2 (attitude and intention)

Select just each "batch" batch of questions separtely and add a new column that counts NAs for each question set.

```{r}

perception <- important_data %>%
  select(response_id, group, starts_with("perception")) %>%
   mutate(per_missed = rowSums(is.na(.))) %>%
  mutate(per_exclude = case_when(per_missed == 0 ~ 'Include',
                               TRUE ~ 'Exclude'))

attitude <- important_data %>%
  select(response_id, starts_with("attitude")) %>%
   mutate(att_missed = rowSums(is.na(.))) %>%
  mutate(att_exclude = case_when(att_missed <= 2 ~ 'Include',
                               TRUE ~ 'Exclude'))

intention <- important_data %>%
  select(response_id, starts_with("intention")) %>%
   mutate(int_missed = rowSums(is.na(.))) %>%
   mutate(int_exclude = case_when(int_missed <= 2 ~ 'Include',
                               TRUE ~ 'Exclude'))

manipcheck <- important_data %>%
  select(manipulation_check_gf_1, manipulation_check_lf_1)

all_ratings <- cbind(perception, attitude, intention, manipcheck) %>%
  select(-6, -14) %>% # drop extra id columns
 relocate(per_missed, .before = int_missed) %>%
   relocate(att_missed, .before = int_missed) %>%
  relocate(per_exclude, .before = int_exclude) %>%
  relocate(att_exclude, .before = int_exclude) 

```

count how many participants in question type to exclude?
```{r}

all_ratings %>%
  tabyl(per_exclude)

all_ratings %>%
  tabyl(att_exclude)

all_ratings %>%
  tabyl(int_exclude)

```

Look at each case separately...

Perception first, for this single item it doesn't matter; the analysis will exclude the item because it is NA. 
```{r}
per_ex <- all_ratings %>%
  filter(per_exclude == "Exclude")

```

Intention, also not a problem. One participant missed ALL of the intention questions. Because their responses for all items are already NA can just leave it as is.  Their responses for the overall and analyses questions need to be changed to NA. 

```{r}
int_ex <- all_ratings %>%
  filter(int_exclude == "Exclude")
```

Attitude, this one needs to be fixed. One participant missed design, data, and access questions. Their responses for two questions they did answer (overall analyses) need to be changed to NA. 
```{r}
att_ex <- all_ratings %>%
  filter(att_exclude == "Exclude")

# who is that guy?
att_ex$response_id

```


Switching to ifelse because I couldnt work out how to stop it from turning ALL the values in the col to NA. ifelse works like if responseID == the bad guy, turn it to NA, otherwise put value from att_overall1
```{r}
all_ratings <- all_ratings %>%
  mutate(attitude_overall_1 = ifelse(response_id == "R_1F2IfmtZ4IB7wDR", NA, attitude_overall_1)) %>%
  mutate(attitude_analyses_1 = ifelse(response_id == "R_1F2IfmtZ4IB7wDR", NA, attitude_analyses_1)) 
  
# check that problem is fixed

all_ratings %>%
  filter(response_id == "R_1F2IfmtZ4IB7wDR")

# yes

```

# NEXT calculate row means.... 

```{r}

mean_ratings <- all_ratings %>%
  mutate(mean_att = rowMeans(.[4:8], na.rm=TRUE)) %>%
  mutate(mean_int = rowMeans(.[9:12], na.rm=TRUE)) %>%
  select (response_id, group, perception_1, mean_att, mean_int, manipulation_check_gf_1, manipulation_check_lf_1)

```


2.  write clean data to .csv calling it "dateStudy1data_clean.csv"

```{r}

mean_ratings %>%
  write_csv(here("data", "dateStudy1data_clean.csv" ))

```

# demographics

Group

```{r}

data %>%
  count(group)

```

```{r}

important_data %>%
  count(group)

```

Gender

```{r}

data %>%
  count(gender)

```

```{r}

data %>%
  group_by(group) %>%
  count(gender)

```

Age

```{r}

data %>%
  count(age_in_years)

```

```{r}

data %>%
  group_by(group) %>%
  count(age_in_years)

```

Faculty of study

```{r}

data %>%
  count(faculty_of_study)

```

```{r}

data %>%
  group_by(group) %>%
  count(faculty_of_study)

```

Research experience

```{r}

data %>%
  count(research_experience)

```

```{r}

data %>%
  group_by(group) %>%
  count(research_experience)

```
