==> scripts/3_study1descriptives.Rmd <==
---
title: "3_study1descriptives"
author: "Jen Richmond"
date: "13/07/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# load packages
```{r message=FALSE, warning=FALSE}
library(tidyverse)
library(here)
library(gt)

```

# read data
```{r message=FALSE, warning=FALSE}
clean_data <- read_csv(here("data", "dateStudy1data_clean.csv"))

clean_data <- clean_data %>%
  mutate(group = factor(group, 
                        levels = c(1, 2),
                        labels = c("gain", "loss")))

```

Categorical variable: group
Continuous variable: perception_1, mean_att, mean_int

# manipulation check

Mean ratings for gain and loss manipulation check by group

```{r}

summary_manipcheck <- clean_data %>%
  group_by(group) %>%
  summarise(gf_M = mean(manipulation_check_gf_1, na.rm = TRUE),
            gfSD = sd(manipulation_check_gf_1, na.rm = TRUE), 
            gfN = n(), 
            gfStdErr = gfSD/sqrt(gfN),
            lf_M = mean(manipulation_check_lf_1, na.rm = TRUE),
            lfSD = sd(manipulation_check_lf_1, na.rm = TRUE), 
            lfN = n(), 
            lfStdErr =  lfSD/sqrt(lfN))

print(summary_manipcheck)
  
  # the gt() package will give you nicely formatted tables in your knitted documnet, you can also control digits displayed with fmt_number
  
    summary_manipcheck %>% 
      gt() %>%
    fmt_number(columns = c("gf_M", "gfSD", "gfStdErr", "lf_M", "lfSD", "lfStdErr"), 
               decimals = 4)

```

Mean ratings for perception by group

```{r}

summary_perception <- clean_data %>%
  group_by(group) %>%
  summarise(mean = mean(perception_1, na.rm=TRUE), 
            sd = sd(perception_1, na.rm=TRUE), 
            n = n(), 
            stderr = sd / sqrt(n))

  print(summary_perception)

```

Mean ratings for attitude by group

```{r}

summary_attitude <- clean_data %>%
  group_by(group) %>%
  summarise(mean = mean(mean_att, na.rm=TRUE), 
            sd = sd(mean_att, na.rm=TRUE), 
            n = n(), 
            stderr = sd / sqrt(n))

  print(summary_attitude)

```

Mean ratings for intention by group

```{r}

summary_intention <- clean_data %>%
  group_by(group) %>%
  summarise(mean = mean(mean_int, na.rm=TRUE), 
            sd = sd(mean_int, na.rm=TRUE), 
            n = n(), 
            stderr = sd / sqrt(n))

  print(summary_intention)

```

All mean ratings

```{r}

summary_data <- clean_data %>%    
  group_by(group) %>%
  summarise(gf_M = mean(manipulation_check_gf_1, na.rm = TRUE),
            gfSD = sd(manipulation_check_gf_1, na.rm = TRUE), 
            gfN = n(), 
            gfStdErr = gfSD/sqrt(gfN),
            lf_M = mean(manipulation_check_lf_1, na.rm = TRUE),
            lfSD = sd(manipulation_check_lf_1, na.rm = TRUE), 
            lfN = n(), 
            lfStdErr =  lfSD/sqrt(lfN),
            per_M = mean(perception_1, na.rm = TRUE),
            perSD = sd(perception_1, na.rm = TRUE), 
            perN = n(), 
            perStdErr =  perSD/sqrt(perN),
            att_M = mean(mean_att, na.rm = TRUE),
            attSD = sd(mean_att, na.rm = TRUE), 
            attN = n(), 
            attStdErr =  attSD/sqrt(attN),
            int_M = mean(mean_att, na.rm = TRUE),
            intSD = sd(mean_att, na.rm = TRUE), 
            intN = n(), 
            intStdErr =  intSD/sqrt(intN))

print(summary_data)

```

Saving summary data (M, SD, N, StdErr) as a csv

```{r}

summary_data %>%
  write_csv(here("data", "summary_data.csv" ))

```


# TIPS FROM JENNY

I wonder whether there is a way to get these summary stats using less repetitive code. I think if we make the data into LONG format, we can get them using group_by and summarise all in one go (rather writing code for perception, attitude and intention separately)

Lets rename the columns first...

```{r}
clean_data <- clean_data %>%
  rename(perception = perception_1, attitude = mean_att, intention = mean_int, positive_benefits = manipulation_check_gf_1, negative_conseq = manipulation_check_lf_1)

names(clean_data)
```

Then make the dataframe long, using pivot longer...

```{r}
clean_data_long <- clean_data %>%
  pivot_longer(names_to = "question", values_to = "rating", perception:negative_conseq)
```

Then do summary stats, manipulation check first...

```{r}
manipcheck <- clean_data_long %>%
  filter(question %in% c("positive_benefits", "negative_conseq")) %>%
  group_by(group, question) %>%
  summarise(mean = mean(rating, na.rm = TRUE),
            SD = sd(rating, na.rm = TRUE), 
            N = n(), 
            StdErr = SD/sqrt(N)) 

manipcheck %>%
  gt() %>%
  fmt_number(columns = c("mean", "SD", "StdErr"), 
               decimals = 4)
```

Now do the same for perception, attitude, intention....

```{r}

per_att_int <- clean_data_long %>%
  filter(question %in% c("perception", "attitude", "intention")) %>%
  group_by(group, question) %>%
  summarise(mean = mean(rating, na.rm = TRUE),
            SD = sd(rating, na.rm = TRUE), 
            N = n(), 
            StdErr = SD/sqrt(N)) 

per_att_int %>%
  gt() %>%
  fmt_number(columns = c("mean", "SD", "StdErr"), 
               decimals = 4)

```

```{r}

manipcheck %>%
  write_csv(here("data", "manipcheck.csv" ))

per_att_int %>%
  write_csv(here("data", "per_att_int.csv" ))

```
