==> scripts/4_study1plots.Rmd <==
---
title: "4_study1plots"
author: "Jen Richmond"
date: "13/07/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# load packages
 
```{r message=FALSE, warning=FALSE}
library(tidyverse)
library(here)
library(ggeasy)

```

# read data

```{r message=FALSE, warning=FALSE}
clean_data <- read_csv(here("data", "dateStudy1data_clean.csv"))

summary_data <- read_csv(here("data", "summary_data.csv"))

manipcheck <- read_csv(here("data", "manipcheck.csv"))

per_att_int <- read_csv(here("data", "per_att_int.csv"))

```

Categorical variable: group
Continuous variable: perception_1, mean_att, mean_int

#Mean ratings for gain manipulation check by group

Column graph:

```{r}

gainmanip <- manipcheck %>%
  filter(question == "positive_benefits") 

gainmanip %>%
    ggplot(aes(x = group, y = mean, fill = group)) +
    geom_col(width = 0.5, show.legend = FALSE) +
    scale_y_continuous(expand = c(0,0),
                       limits = c(0,5)) +
    geom_errorbar(aes(ymin=mean-StdErr, ymax=mean+StdErr),
                  size=.3, 
                  width=.2) +
  theme_classic() +
  labs(title = "To what extent did the information stress the positive benefits of \nadopting open science practices?", x="Condition", y="Mean ratings of gain manipulation")

```


Geom jitter plot:

> JR note you only need to do this once, once you have converted group in the clean_data df to a factor and told R you want the levels to be listed as Gain Loss, it will use that same factor structure every time you use the dataframe. 

```{r}

clean_data <- clean_data %>%
  mutate(group = factor(group, 
                        levels = c(1, 2),
                        labels = c("Gain", "Loss")))

```

```{r}

clean_data %>%
  group_by(group) %>%
  ggplot(aes(x = group, y = manipulation_check_gf_1, colour = group)) +
  geom_jitter(width = 0.3, show.legend = FALSE) +
  scale_y_continuous(expand = c(0,0),
                     limits = c(0,6)) +
  theme_classic() +
  labs(title = "To what extent did the information stress the positive benefits of \nadopting open science practices?", x="Condition", y="Mean ratings of gain manipulation")

```


#Mean ratings for loss manipulation check by group

Column graph:



```{r}

lossmanip <- manipcheck %>%
  filter(question == "negative_conseq")

lossmanip %>%
    ggplot(aes(x = group, y = mean, fill = group)) +
    geom_col(width = 0.5, show.legend = FALSE) +
    scale_y_continuous(expand = c(0,0),
                       limits = c(0,5)) +
    geom_errorbar(aes(ymin=mean-StdErr, ymax=mean+StdErr),
                  size=.3, 
                  width=.2) +
  theme_classic() +
  labs(title = "To what extent did the information stress the negative consequences of \nnot adopting open science practices?", x="Condition", y="Mean ratings of loss manipulation")

```


Geom jitter plot:


```{r}

glimpse(clean_data)

clean_data %>%
  group_by(group) %>%
  ggplot(aes(x = group, y = manipulation_check_lf_1, colour = group)) +
  geom_jitter(width = 0.3, show.legend = FALSE) +
  scale_y_continuous(expand = c(0,0),
                     limits = c(0,6)) +
  theme_classic() +
  labs(title = "To what extent did the information stress the negative consequences of \nnot adopting open science practices?", x="Condition", y="Mean ratings of loss manipulation")

```



# Mean ratings for gain and loss manipulation check by group

#Facet_wrap()


> JC. Is there a way to rename "gain" to "Gain", "loss" to "Loss", "gainmanipulation" to "Gain Manipulation", and "lossmanipulation" to "Loss Manipulation"? I tried using the rename() function but can only seem to do it for variable names. I also tried levels(manipcheck$group) <- c("gain to Gain") but it didn't work. Would it be mutating the labels? i can't work out the exact function

> JR yes see factors method below

```{r}

glimpse(manipcheck) # check what kind of data R thinks you have = char
  
# alternative renaming method converting to factors and changing labels
  
manipcheck<- manipcheck %>%
  mutate(group = factor(group, 
                        levels = c("gain", "loss"),
                        labels = c("Gain", "Loss")))

levels(manipcheck$group) # check the levels of the new factor

manipcheck<- manipcheck %>%
  mutate(question = factor(question, 
                        levels = c("negative_conseq", "positive_benefits"),
                        labels = c("Negative consequences", "Positive benefits")))

levels(manipcheck$group) # check the levels of the new factor
levels(manipcheck$question)
  
```
> JR Ive left the question for the moment, it is listed as positive benefits vs negative conseq which I think is more meaningful than gain and loss manipulation (which conflates it with group and makes it confusing to me)

```{r}

manipcheck %>%
    ggplot(aes(x = group, y = mean, fill = group)) +
    geom_col(width = 0.7, position = "dodge", show.legend = FALSE) +
    scale_y_continuous(expand = c(0,0),
                       limits = c(0,5)) +
    theme_classic() +
    facet_wrap(~ question) +
    geom_errorbar(aes(ymin=mean-StdErr, ymax=mean+StdErr),
                  size=.3,
                  width=.2,
                  position=position_dodge(.9)) +
     labs(x="Condition", y="Mean ratings of information")

```

```{r}

manipcheck <- manipcheck %>%
  mutate(group = factor(group, 
                        levels = c("gain", "loss"),
                        labels = c("Gain group", "Loss group")))

manipcheck <- manipcheck %>%
  mutate(question = factor(question, 
                        levels = c("positive_benefits", "negative_conseq"),
                        labels = c("Positive benefits", "Negative consequences")))

manipcheck %>%
    ggplot(aes(x = question, y = mean, fill = question)) +
    geom_col(width = 0.7, position = "dodge", show.legend = FALSE) +
    scale_y_continuous(expand = c(0,0),
                       limits = c(0,5)) +
    theme_classic() +
    facet_wrap(~ group) +
    geom_errorbar(aes(ymin=mean-StdErr, ymax=mean+StdErr),
                  size=.3,
                  width=.2,
                  position=position_dodge(.9)) +
     labs(x="Type of information", y="Mean ratings of information")

```

#Mean ratings for perception by group

Column graph:
> JR- note it might be simpler to filter straight into your ggplots code. Rather than creating a new df that is filtered, you can pipe from filter stragith into ggplot 

```{r}

per_att_int %>%
  filter(question == "perception") %>%
  ggplot(aes(x = group, y = mean, fill = group)) +
  geom_col(width = 0.5, show.legend = FALSE) +
  scale_y_continuous(expand = c(0,0),
                     limits = c(0,5)) +
  geom_errorbar(aes(ymin=mean-StdErr, ymax=mean+StdErr),
                size=.3, 
                width=.2) +
  theme_classic() +
  labs(x="Condition", y="Mean ratings of perception")
  
```

Geom jitter plot:

```{r}


clean_data %>%
  group_by(group) %>%
  ggplot(aes(x = group, y = perception_1, colour = group)) +
  geom_jitter(width = 0.3, show.legend = FALSE) +
  scale_y_continuous(expand = c(0,0),
                     limits = c(0,6)) +
  theme_classic() +
  labs(x="Condition", y="Mean ratings of perception")

```


#Mean ratings for attitude by group

Column graph:

```{r}

per_att_int %>%
  filter(question == "attitude") %>%
  ggplot(aes(x = group, y = mean, fill = group)) +
  geom_col(width = 0.5, show.legend = FALSE) +
  scale_y_continuous(expand = c(0,0),
                     limits = c(0,5)) +
  geom_errorbar(aes(ymin=mean-StdErr, ymax=mean+StdErr),
                size=.3, 
                width=.2) +
  theme_classic() +
  labs(x="Condition", y="Mean ratings of attitude")

```

Geom jitter plot:

```{r}


clean_data %>%
  group_by(group) %>%
  ggplot(aes(x = group, y = mean_att, colour = group)) +
  geom_jitter(width = 0.3, show.legend = FALSE) +
  scale_y_continuous(expand = c(0,0),
                     limits = c(0,6)) +
  theme_classic() +
  labs(x="Condition", y="Mean ratings of attitude")

```

#Mean ratings for intention by group

Column graph:



```{r}

per_att_int %>%
  filter(question == "intention") %>%
  ggplot(aes(x = group, y = mean, fill = group)) +
  geom_col(width = 0.5, show.legend = FALSE) +
  scale_y_continuous(expand = c(0,0),
                     limits = c(0,5)) +
  geom_errorbar(aes(ymin=mean-StdErr, ymax=mean+StdErr),
                size=.3, 
                width=.2) +
  theme_classic() +
  labs(x="Condition", y="Mean ratings of intention")

```

Geom jitter plot:

```{r}


clean_data %>%
  group_by(group) %>%
  ggplot(aes(x = group, y = mean_int, colour = group)) +
  geom_jitter(width = 0.3, show.legend = FALSE) +
  scale_y_continuous(expand = c(0,0),
                     limits = c(0,6)) +
  theme_classic() +
  labs(x="Condition", y="Mean ratings of intention")

```
#Mean ratings for perception, attitude, and intention by group

#Facet_wrap()

> JR this method works, but the code is a bit hard to see what it is doing. Can you use the same factor method above to convert question to factor and set the order??

```{r}

glimpse(per_att_int)

  per_att_int$group[per_att_int$group == "gain"] <- "Gain"
  per_att_int$group[per_att_int$group == "loss"] <- "Loss"
  per_att_int$question[per_att_int$question == "perception"] <- "a) Perception"
  per_att_int$question[per_att_int$question == "attitude"] <- "b) Attitude"
  per_att_int$question[per_att_int$question == "intention"] <- "c) Intention"

per_att_int_ordered <- per_att_int
per_att_int_ordered$question <- factor(per_att_int_ordered$question,
                         levels = c("a) Perception", "b) Attitude", "c) Intention"))

per_att_int_ordered %>%
    ggplot(aes(x = group, y = mean, fill = group)) +
    geom_col(width = 0.7, position = "dodge", show.legend = FALSE) +
    scale_y_continuous(expand = c(0,0),
                       limits = c(0,5)) +
    theme_classic() +
    facet_wrap(~ question) +
    geom_errorbar(aes(ymin=mean-StdErr, ymax=mean+StdErr),
                  size=.3,
                  width=.2,
                  position=position_dodge(.9)) +
     labs(x="Condition", y="Mean ratings")

```

#Likert

```{r}

library(likert)

```
