==> scripts/5_study1stats.Rmd <==
---
title: "5_study1stats"
author: "Jen Richmond"
date: "13/07/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(tidyverse)
library(here)
library(jmv)
```


```{r}
clean_data <- read_csv(here("data", "dateStudy1data_clean.csv"))
```

Gain-Frame Manipulation Check

```{r}

clean_data %>%
ttestIS(formula = manipulation_check_gf_1 ~ group)

```

Loss-Frame Manipulation Check

```{r}

clean_data %>%
ttestIS(formula = manipulation_check_lf_1 ~ group)

```

Perception

```{r}

clean_data %>%
ttestIS(formula = perception_1 ~ group)

```

Attitude

```{r}

clean_data %>%
ttestIS(formula = mean_att ~ group)

```

Intention

```{r}

clean_data %>%
ttestIS(formula = mean_int ~ group)

```
