==> scripts/7_cleanStudy2_Qualtricsdata.Rmd <==
---
title: "Clean Qualtrics Study 2 Data"
author: "Jenny Chen"
date: "11/08/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Loading packages

```{r message=FALSE, warning=FALSE}

library(tidyverse)
library(here)
library(janitor)

```

```{r message=FALSE, warning=FALSE}

datatest <-  read_csv(here("Study_2_Data.csv"))

```

# Selecting just variables of interest

```{r}

data_select <- datatest %>%
  select(EndDate, Progress, ResponseId, Group, Gender, "Age in Years", "Faculty of Study", "Research Experience", Perception_1, starts_with("Attitude"), starts_with("Intention"), "ManipulationCheck GF_1", "ManipulationCheck LF_1", starts_with("AttentionAttempt"),  starts_with("Score"), "Response Check_1")

```

# Filtering out participants who did not complete the whole survey

212 --> 200 participants = 8 participants did not complete the whole survey

```{r}

progress_filtered <- data_select %>%
  filter(Progress == "100")

progress_filtered %>% 
  count()

```

# Creating new exclude column for attention scores

```{r}

progress_filtered <- progress_filtered %>%
  rename(score_attempt_1 = "Score Attempt 1") %>%
  rename(score_attempt_2 = "Score Attempt 2") %>%
  rename(score_attempt_3 = "Score Attempt 3")

```

```{r}

exclude_column <- progress_filtered %>%
  mutate(atten_exclusion = case_when(score_attempt_1 >= 4 ~ 'Include',
                               score_attempt_2 >= 4 ~ 'Include',
                               score_attempt_3 >= 4 ~ 'Include',
                               TRUE ~ 'Exclude'))

```

# Filtering out those that get exclude in the above column (failed attention check)

> Most of the participants who did not pass the attention check also failed the response bias check that I excluded above. Is this ok when reporting in the results? "A further ___ were excluded from failing the attention check". Or do I have to give the exact number of exclusions for progress, response bias, and attention check, even though there is a lot of overlap

> JR I think you shouldn't be thinking of these as overlapping... I think there are levels of not being inlcuded, if you didn't complete you are out... Out of the people who completed, the next hurdle was to pass the attention check... and THEN out of people who 1. completed, and 2. learned somethnig form the information, we want to throw out those who were probably answering questions randomly... 

200 --> 189 participants = 11 participants failed attention check

```{r}

attention_filtered <- exclude_column %>%
  filter(atten_exclusion == "Include")

attention_filtered %>% 
  count()

```

# Excluding participants who did not answer the response bias check accurately

> Problem: didn't realise leaving the slider at '3' would result in NA
Change NAs to 3s? Safe to assume people who didn't get excluded from attention check or missing Qs actually answered '3' instead of skipping the question?

189 --> 180 participants = 9 participants did not pass the response bias check

```{r}

renamed_attention <- attention_filtered %>%
  rename(response_check_1 = "Response Check_1")

```

```{r}

responsecheck_filtered <- renamed_attention %>%
  filter(is.na(response_check_1) | response_check_1 >= 2.75 & response_check_1 <= 3.25) 

responsecheck_filtered %>% 
  count()

important_data <- responsecheck_filtered %>%
  select(EndDate, Progress, ResponseId, Group, Gender, "Age in Years", "Faculty of Study", "Research Experience", Perception_1, starts_with("Attitude"), starts_with("Intention"), "ManipulationCheck GF_1", "ManipulationCheck LF_1", atten_exclusion, response_check_1)
  
```

# Excluding participants who did not answer enough perception/attitude/intention questions

From pre-registration: If a participant does not complete the single perceived severity question, at least 2 of 4 attitude questions, or at least 2 of 3 implementation intention questions, the participant’s response will be excluded from the analysis for that particular dependent variable.

```{r}

perception <- important_data %>%
  select(ResponseId, Group, starts_with("Perception")) %>%
  mutate(per_missed = rowSums(is.na(.))) %>%
  mutate(per_exclude = case_when(per_missed == 0 ~ 'Include',
                               TRUE ~ 'Exclude'))

attitude <- important_data %>%
  select(ResponseId, starts_with("Attitude")) %>%
  mutate(att_missed = rowSums(is.na(.))) %>%
  mutate(att_exclude = case_when(att_missed <= 2 ~ 'Include',
                               TRUE ~ 'Exclude'))

intention <- important_data %>%
  select(ResponseId, starts_with("Intention")) %>%
  mutate(int_missed = rowSums(is.na(.))) %>%
  mutate(int_exclude = case_when(int_missed <= 1 ~ 'Include',
                               TRUE ~ 'Exclude'))

manipcheck <- important_data %>%
  select("ManipulationCheck GF_1", "ManipulationCheck LF_1")

all_ratings <- cbind(perception, attitude, intention, manipcheck) %>%
  select(-6, -13) %>% # dropping extra response_id columns
  relocate(per_missed, .before = int_missed) %>%
  relocate(att_missed, .before = int_missed) %>%
  relocate(per_exclude, .before = int_exclude) %>%
  relocate(att_exclude, .before = int_exclude) 

```

# Counting how many participants in each question type to exclude

```{r}

all_ratings %>%
  tabyl(per_exclude)

all_ratings %>%
  tabyl(att_exclude)

all_ratings %>%
  tabyl(int_exclude)

```

```{r}

per_ex <- all_ratings %>%
  filter(per_exclude == "Exclude")

per_ex$response_id

```

```{r}

att_ex <- all_ratings %>%
  filter(att_exclude == "Exclude")

att_ex$ResponseId

```

```{r}

int_ex <- all_ratings %>%
  filter(int_exclude == "Exclude")

int_ex$ResponseId

```

```{r}

all_ratings <- all_ratings %>%
  rename(response_id = "ResponseId")

```

```{r}

all_ratings <- all_ratings %>%
  rename(attitude_access_1 = "Attitude Access_1")

```

```{r}

all_ratings <- all_ratings %>%
  rename(intention_access_1 = "Intention Access_1")

```

```{r}

all_ratings <- all_ratings %>%
  mutate(attitude_access_1 = ifelse(response_id == "R_3EB8OxsQDNpBX1C", NA, attitude_access_1)) 
  
all_ratings %>%
  filter(response_id == "R_3EB8OxsQDNpBX1C")

```

```{r}

all_ratings <- all_ratings %>%
  mutate(intention_access_1 = ifelse(response_id == "R_2dzke7muHlM004v", NA, intention_access_1)) 
  
all_ratings %>%
  filter(response_id == "R_2dzke7muHlM004v")

```
  
# Calculating row means

```{r}

all_ratings %>%
  glimpse()

```

```{r}

all_ratings$Perception_1 <- as.numeric(all_ratings$Perception_1)
all_ratings$`Attitude Overall_1` <- as.numeric(all_ratings$`Attitude Overall_1`)
all_ratings$`Attitude Design_1` <- as.numeric(all_ratings$`Attitude Design_1`)
all_ratings$`Attitude Data&Analys_1` <- as.numeric(all_ratings$`Attitude Data&Analys_1`)
all_ratings$attitude_access_1 <- as.numeric(all_ratings$attitude_access_1)
all_ratings$`Intention Design_1` <- as.numeric(all_ratings$`Intention Design_1`)
all_ratings$`Intention Data_1` <- as.numeric(all_ratings$`Intention Data_1`)
all_ratings$intention_access_1 <- as.numeric(all_ratings$intention_access_1)

```

```{r}

mean_ratings <- all_ratings %>%
  mutate(mean_att = rowMeans(.[4:7], na.rm = TRUE)) %>%
  mutate(mean_int = rowMeans(.[8:10], na.rm = TRUE)) %>%
  select(response_id, Group, Perception_1, mean_att, mean_int, "ManipulationCheck GF_1", "ManipulationCheck LF_1")

```

 mean_ratings %>%
  tabyl(group)

# Writing clean data to .csv calling it "dateStudy2data_clean.csv"

```{r}

mean_ratings %>%
  write_csv(here("data", "dateStudy2data_clean.csv" ))

```

# demographics

Group

```{r}

datatest %>%
  count(Group)

```

```{r}

important_data %>%
  count(Group)

```

Gender

```{r}

datatest %>%
  count(Gender)

```

```{r}

datatest %>%
  group_by(Group) %>%
  count(Gender)

```

Age
  
```{r}

datatest <- datatest %>%
  rename(age_in_years = "Age in Years")
  
```
  
```{r}

datatest %>%
  count(age_in_years)

```

```{r}

datatest %>%
  group_by(Group) %>%
  count(age_in_years)

```

Faculty of study

```{r}

datatest <- datatest %>%
  rename(faculty_of_study = "Faculty of Study")
  
```

```{r}

datatest %>%
  count(faculty_of_study)

```

```{r}

datatest %>%
  group_by(Group) %>%
  count(faculty_of_study)

```

Research experience

```{r}

datatest <- datatest %>%
  rename(research_experience = "Research Experience")
  
```

```{r}

datatest %>%
  count(research_experience)

```

```{r}

datatest %>%
  group_by(Group) %>%
  count(research_experience)

```
