==> scripts/10_study2stats.Rmd <==
---
title: "Study 2 Statistics"
author: "Jenny Chen"
date: "11/08/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Loading packages

```{r}

library(tidyverse)
library(here)
library(jmv)

```


```{r}

clean_data <- read_csv(here("data", "dateStudy2data_clean.csv"))

```

# Performing independent t-tests

# Gain-Frame Manipulation Check

```{r}

clean_data %>%
ttestIS(formula = `ManipulationCheck GF_1` ~ Group)

```

# Loss-Frame Manipulation Check

```{r}

clean_data %>%
ttestIS(formula = `ManipulationCheck LF_1` ~ Group)

```

# Perception

```{r}

clean_data %>%
ttestIS(formula = Perception_1 ~ Group)

```

# Attitude

```{r}

clean_data %>%
ttestIS(formula = mean_att ~ Group)

```

# Intention

```{r}

clean_data %>%
ttestIS(formula = mean_int ~ Group)

```
